Avoid some warnings when compiling with windows
git-svn-id: https://deskvox.svn.sourceforge.net/svnroot/deskvox/trunk@449 04231f92-3938-0410-9931-e931ac552b4f

+++ virvo/src/vvbsptree.cpp
@@ -117,20 +117,20 @@ vvRect* vvAABB::getProjectedScreenRect()
 
     if (vertex[0] < minX)
     {
-      minX = vertex[0];
+      minX = static_cast<float>(vertex[0]);
     }
     if (vertex[0] > maxX)
     {
-      maxX = vertex[0];
+      maxX = static_cast<float>(vertex[0]);
     }
 
     if (vertex[1] < minY)
     {
-      minY = vertex[1];
+      minY = static_cast<float>(vertex[1]);
     }
     if (vertex[1] > maxY)
     {
-      maxY = vertex[1];
+      maxY = static_cast<float>(vertex[1]);
     }
   }
